Fix overloaded funtion to call I2C master begin

+++ src/Hat_PowerC.cpp
@@ -7,7 +7,7 @@ bool PowerC::begin(TwoWire *wire, uint8_t sda, uint8_t scl, uint8_t addr) {
     _addr = addr;
     _sda  = sda;
     _scl  = scl;
-    _wire->begin(_sda, _scl, 400000UL);
+    _wire->begin((int)_sda, (int)_scl, 400000UL);
     _wire->beginTransmission(_addr);
     uint8_t error = _wire->endTransmission();
     if (error == 0) {
